+++ app/charts/scatterplot/chart-scatterplot.tsx
@@ -65,7 +65,7 @@ const ChartScatterplot = memo((props: ChartProps<ScatterPlotConfig>) => {
               dimensionsById={dimensionsById}
               chartConfig={chartConfig}
               symbol="circle"
-              interactive={interactiveFiltersConfig?.legend.active}
+              interactive={interactiveFiltersConfig.legend.active}
               showTitle={fields.segment.showTitle}
             />
           )}