+++ app/configurator/interactive-filters/interactive-filters-config-state.tsx
@@ -1,3 +1,4 @@
+import { SelectChangeEvent } from "@mui/material";
 import produce from "immer";
 import get from "lodash/get";
 import { ChangeEvent } from "react";
@@ -8,17 +9,18 @@ import {
   isConfiguring,
   useConfiguratorState,
 } from "@/configurator/configurator-state";
+import { FIELD_VALUE_NONE } from "@/configurator/constants";
 import { useEvent } from "@/utils/use-event";
 
-export const useInteractiveFiltersToggle = (target: "legend") => {
+export const useInteractiveFiltersToggle = () => {
   const [state, dispatch] = useConfiguratorState(isConfiguring);
   const chartConfig = getChartConfig(state);
   const onChange = useEvent((e: ChangeEvent<HTMLInputElement>) => {
-    if (chartConfig.interactiveFiltersConfig?.[target]) {
+    if (chartConfig.interactiveFiltersConfig.legend) {
       const newConfig = produce(
         chartConfig.interactiveFiltersConfig,
         (draft) => {
-          draft[target].active = e.currentTarget.checked;
+          draft.legend.active = e.currentTarget.checked;
         }
       );
 
@@ -29,14 +31,11 @@ export const useInteractiveFiltersToggle = (target: "legend") => {
     }
   });
 
-  const stateValue = get(
-    chartConfig,
-    `interactiveFiltersConfig.${target}.active`
-  );
+  const stateValue = get(chartConfig, "interactiveFiltersConfig.legend.active");
   const checked = stateValue ? stateValue : false;
 
   return {
-    name: target,
+    name: "legend",
     checked,
     onChange,
   };
@@ -48,7 +47,7 @@ export const useInteractiveFiltersToggle = (target: "legend") => {
 export const useInteractiveDataFilterToggle = (dimensionId: string) => {
   const [state, dispatch] = useConfiguratorState(isConfiguring);
   const chartConfig = getChartConfig(state);
-  const toggle = useEvent(() => {
+  const onChange = useEvent(() => {
     const { interactiveFiltersConfig } = chartConfig;
     const newIFConfig = toggleInteractiveFilterDataDimension(
       interactiveFiltersConfig,
@@ -61,11 +60,15 @@ export const useInteractiveDataFilterToggle = (dimensionId: string) => {
     });
   });
   const checked =
-    chartConfig.interactiveFiltersConfig?.dataFilters.componentIds?.includes(
+    chartConfig.interactiveFiltersConfig.dataFilters.componentIds.includes(
       dimensionId
     );
 
-  return { checked, toggle };
+  return {
+    name: "dataFilters",
+    checked,
+    onChange,
+  };
 };
 
 // Add or remove a dimension from the interactive  data filters dimensions list
@@ -75,20 +78,26 @@ export const toggleInteractiveFilterDataDimension = produce(
     id: string,
     newValue?: boolean
   ): InteractiveFiltersConfig => {
-    if (!config?.dataFilters.componentIds) {
-      return config;
-    }
-
     const currentComponentIds = config.dataFilters.componentIds;
     const shouldAdd =
       newValue === undefined ? !currentComponentIds.includes(id) : newValue;
     const newComponentIds = shouldAdd
       ? [...currentComponentIds, id]
       : config.dataFilters.componentIds.filter((d) => d !== id);
+
+    const newDefaultValueOverrides = {
+      ...config.dataFilters.defaultValueOverrides,
+    };
+
+    if (!shouldAdd) {
+      delete newDefaultValueOverrides[id];
+    }
+
     const newDataFilters: NonNullable<InteractiveFiltersConfig>["dataFilters"] =
       {
         ...config.dataFilters,
         componentIds: newComponentIds,
+        defaultValueOverrides: newDefaultValueOverrides,
       };
     newDataFilters.active = newComponentIds.length > 0;
 
@@ -116,7 +125,7 @@ export const useInteractiveTimeRangeToggle = () => {
       value: newIFConfig,
     });
   });
-  const checked = chartConfig.interactiveFiltersConfig?.timeRange.active;
+  const checked = chartConfig.interactiveFiltersConfig.timeRange.active;
 
   return { checked, toggle };
 };
@@ -132,4 +141,34 @@ const toggleInteractiveTimeRangeFilter = (config: InteractiveFiltersConfig) => {
   };
 
   return { ...config, timeRange: newTimeRange };
+};
+
+export const useDefaultValueOverride = (dimensionId: string) => {
+  const [state, dispatch] = useConfiguratorState(isConfiguring);
+  const chartConfig = getChartConfig(state);
+  const currentValue =
+    chartConfig.interactiveFiltersConfig.dataFilters.defaultValueOverrides[
+      dimensionId
+    ] || FIELD_VALUE_NONE;
+
+  const onChange = useEvent((e: SelectChangeEvent<unknown>) => {
+    const newValue = e.target.value as string;
+    const newConfig = produce(chartConfig.interactiveFiltersConfig, (draft) => {
+      if (newValue === FIELD_VALUE_NONE) {
+        delete draft.dataFilters.defaultValueOverrides[dimensionId];
+      } else {
+        draft.dataFilters.defaultValueOverrides[dimensionId] = newValue;
+      }
+    });
+
+    dispatch({
+      type: "INTERACTIVE_FILTER_CHANGED",
+      value: newConfig,
+    });
+  });
+
+  return {
+    value: currentValue,
+    onChange,
+  };
 };